llm_customer_service/requirements-atguigu.txt
# atguigu_ai 依赖列表
# 此文件由 setup.py 加载，用于 pip install 安装

# ============================================================
# 核心依赖
# ============================================================

# CLI 工具
click>=8.0.0,<9.0.0

# 环境变量管理
python-dotenv>=1.0.0,<2.0.0

# Web API 框架
fastapi>=0.100.0,<1.0.0
uvicorn>=0.20.0,<1.0.0

# 数据模型与验证
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0

# ============================================================
# LLM 集成
# ============================================================

# LangChain 核心
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0

# LangChain 模型提供商
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0

# LangGraph 对话流程引擎
langgraph>=0.2.0

# OpenAI API
openai>=1.0.0,<2.0.0

# 阿里云 DashScope (通义千问)
dashscope>=1.20.0

# ============================================================
# 嵌入与检索
# ============================================================

# PyTorch (sentence-transformers 依赖)
# 注意：torch 安装较大，如果不使用本地嵌入模型可以注释掉
torch>=2.0.0

# 本地嵌入模型
sentence-transformers>=2.2.0

# 向量计算
numpy>=1.24.0,<3.0.0

# ============================================================
# 数据存储
# ============================================================

# ORM 框架
sqlalchemy>=2.0.0,<3.0.0

# MySQL 驱动（可选：使用MySQL存储时需要）
pymysql>=1.0.0

# Neo4j 图数据库（GraphRAG 知识图谱检索需要）
neo4j>=5.0.0
neo4j-graphrag>=1.0.0

# ============================================================
# 中文处理
# ============================================================

# 中文分词（GraphRAG检索需要）
jieba>=0.42.0

# ============================================================
# 配置与模板
# ============================================================

# YAML 处理
pyyaml>=6.0.0,<7.0.0
ruamel.yaml>=0.17.0,<1.0.0

# 模板引擎
jinja2>=3.0.0,<4.0.0

# ============================================================
# 网络通信
# ============================================================

# HTTP 客户端
aiohttp>=3.8.0,<4.0.0
httpx>=0.24.0,<1.0.0

# WebSocket
python-socketio>=5.0.0,<6.0.0

# ============================================================
# 工具库
# ============================================================

# 进度条
tqdm>=4.60.0,<5.0.0

# 终端美化
colorama>=0.4.0,<1.0.0
rich>=10.0.0,<15.0.0

# 类型扩展
typing-extensions>=4.0.0,<5.0.0

# ============================================================
# 开发与测试工具
# ============================================================

# 测试数据生成
faker>=18.0.0

# 数据库工具
dbutils>=3.0.0
sqlacodegen>=3.0.0
